About time I got things under (version) control!
It's nice finally getting to write some assembly :-)

Normally, I'd commit as soon as I've scaffolded an empty repo -
and then commit way too often. This time, I put off the `git init`
until I have a working PoC. This strategy of delaying of the
dopamine hit that comes with each commit has paid off, and
today we have the rudiments of a NEC V53(A?) emulator.

(And when I say rudiments, I mean it - even the flag results of
instructions are not implemented yet.)

The purpose of this thing is to run the OS of the AKAI MPC2000XL
and AKAI S3000XL, in order to enable me to make improvements to
the UX of those devices.

The MAME emulation project contains a working prototype of MPC3K
emulation (as well as MPC60 and a couple more AKAI devices), so I've
used that as a starting point alongside the chip docs.

To get this running, you need to interleave the pair of OS ROM dumps
(which can be found on Archive.org in MAME ROM dump collections),
resulting in a single .bin file that can be loaded into
the emulator's memory.

Once you've fed it that, the emulator will run for a short while,
reach an unimplemented instruction, and crash. The run log looks
quite similar to what you would get in `mame -debug mpc3000`,
so it seems I'm on the right track here :-)

The obvious big milestone is getting as far as writing characters
to the display (thankfully, this looks quite simple: ASCII bytes
are written to an output buffer). That's the extent of the working
emulation in MAME.

Afterwards, I need to figure out how to pass inputs into the device
so I can navigate the menus. I don't know yet if I'll be emulating
the input coprocessor, or just simulating it.

Once I have working display and button input, I'll try loading the
MPC2000XL OS. Breaking it down into functions delineated by the
CALL and RET instructions is the logical next step there.

The big breakthough that I'm aiming for is, of course, **resampling**:
enabling the MPC to record sounds while it's playing a sequence -
and then to map them to pads *without stopping the music*.