Implement INC/DEC, CY flag.
On to implementing flags!

Started by writing part of the some docstrings for the
different behaviors of the flags in the PSW register,
most importantly the CY (carry) flag.

Wrote stub handlers for opcodes 0x40-0x4F (INC, DEC instructions).

This was an example of a "WIP" commit: there are no changes to the
program's runtime behavior just yet; on the other hand, a bunch of
typing has been done towards the implementation of a feature, which
would be a pain to redo.

Then, I amended it to add actual implementations of the INC/DEC
instructions, which was a bunch more copy-pasting and find-replacing.

Now, the MPC3K OS proceeds all the way to address 0xFAC1C (0F E0 11)
which the MAME debugger decodes as BRKXA - Start (or Break?) Extended
Address Mode. This is V53(A)-specific; let's see how it works.