Implement extended addressing; fix stm_w
Apparently, all that "extended addressing" does is
activate an alternate page of memory (0x00000-0xA0000).

I sure as hell wasn't able to find that in the docs, nor
in the MAME NEC emulation code. Thankfully, I could see it
happening in the debugger, so that's what I've done.

Implementing it involved adding some new methods to
`struct CPU`, such as `get_byte`/`set_byte` and `xa`/`set_xa`.

`cpu.address()` was renamed to `cpu.program_address()`. (Is that
the same thing as an instruction pointer?)

`cpu.ea()` was renamed to `cpu.effective_address()`.

I made the memory pages fixed-size, so now the memory image
vec that's passed into `CPU::new` gets copied to the main RAM page.
There's a new `cpu.extended` memory page.

Weirdly, the docs specify that the XAM register is at
"internal I/O address" 0xFF80, but the "internal I/O"
page in MAME is only 256 bytes long and the flag goes in
the "regular" I/O page.

Also fixed (?) the addressing in the `stm_w` instruction
to always be offset from DS1 as per the docs.