Impl BC, BNC; variants of: BR, DIV, CMP, SUB, PUSH
Added `set_register_u8` since it's not possible to obtain
a mutable reference to a byte register (without using union
and unsafe, anyway?).

Added missing `set_pzs` calls to INC/DEC instructions.

The stack stuff around division is weird.

Fixed BE which was flipped.

Next opcode to implement is 0x80 which is another "scary"
new thing (what is IMM even?)