WIP refactor: define_instruction_set -> impl_instruction_set
This will be the second version of the instruction set macro.
It will be more flexible, allowing for instructions to be
defined both in place and as separate functions.

Need to rewrite all the currently implemented instructions
to return closures that take `&mut CPU` and execute the actual
operations. Thus, separate disassembly/execution stages.

OTOH, if returned closures take `&CPU` (not `&mut CPU`), and
only do the read part of the instruction, then return more
closures (this time taking `&mut CPU`) to complete the instruction,
we've got the conceptual foundations of speculative execution or what?
Not very useful when emulating an old chip that doesn't implement
speculative execution, nonetheless it improves my understanding of
how that feature works on a high level.

What's more important is to see whether I can pass around
unboxed closures with `FnOnce` or something.