Fix borrowas
As usual, when the borrow checker acts up,
the solution is to simplify and disentangle.
Who says the instruction dispatcher `fn instruction`
needs to be a method of `struct CPU`, or even a static
method? Make it a free-standing function, pass it a
mutable reference to `CPU` and it will happily return
(and evaluate) closures that take another mutable
reference to `CPU`. Otherwise, the `&mut self` borrow
prevents that.